noGenshinSource/requirements.txt
pillow==9.2.0
wget==3.2
requests==2.27.1